refactor: view controller를 ViewConfig에서 설정
