+++ src/main/java/roomescape/global/config/ViewConfig.java
@@ -1,0 +1,22 @@
+package roomescape.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class ViewConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        // NORMAL USER VIEW
+        registry.addViewController("/reservation").setViewName("reservation");
+        registry.addViewController("/login").setViewName("login");
+
+        // ADMIN USER VIEW
+        registry.addViewController("/admin").setViewName("admin/index");
+        registry.addViewController("/admin/reservation").setViewName("admin/reservation-new");
+        registry.addViewController("/admin/time").setViewName("admin/time");
+        registry.addViewController("/admin/theme").setViewName("admin/theme");
+    }
+}
