+++ src/main/java/roomescape/reservationtime/dto/response/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservationtime.controller;
+package roomescape.reservationtime.dto.response;
 
 import java.time.LocalTime;
 import roomescape.reservationtime.domain.ReservationTime;
